Write playGame() function
To play five rounds of Rock, Paper, Scissors and keep track of the
scores to announce the winner at the end.

I moved playRound() and the score variables to be declared inside
playGame() according to the specifications on The Odin Project. This
also meant reworking the helper functions since they no longer had
access to the previously global score variables. I did this by returning
a variable for whether the human wins, loses, or draws and then updating
the scores within the playRound() function itself. I did not want to
move the helper functions inside the playGame() function since this
would make it harder to read in my opinion.