Create the ui
Used event listeners on buttons for humanChoice and also displayed the
score by changing the text of p elements. This was done to make for a
better user experience.

I also removed all the pseudo code comments as they were cluttering the
file, as well as removing functions no longer needed in the current set
up like getHumanChoice() and getComputerChoiceTest().